requirements.txt
asgiref==3.7.2
Django==4.1
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
python-dotenv==1.0.1
mysqlclient==2.1.1
djangorestframework-dataclasses==1.3.1
faker==22.6.0
pytest==8.0.0
pytest-django==4.8.0